Refactor components and update button styles
- Replaced `<a>` tags with Next.js `<Link>` in `AboutActions.tsx`.
- Updated button variants across multiple components for consistency.
- Introduced new `ProfilePageMainCard` and refactored `ProfilePageContent.tsx`.
- Simplified form handling in `EditPromotions.tsx` and `EditUser.tsx` with new components.
- Removed `data-theme` prop from `Header` in `layout.tsx`.
