+++ Frontend/tkd_hub_web/src/app/components/blogPost/CreatePostButton.tsx
@@ -8,7 +8,7 @@ const CreatePostButton = ({ onClick }: Props) => (
   <div className="mb-6 flex justify-end">
     <GenericButton
       type="button"
-      variant="neutral-dark"
+      variant="primary"
       onClick={onClick}
       className="bg-neutral-900 hover:bg-neutral-800 text-neutral-50"
     >
