+++ Frontend/tkd_hub_web/src/app/components/coaches/CoachModalConfirmDelete.tsx
@@ -29,7 +29,7 @@ const CoachModalConfirmDelete: React.FC<CoachModalConfirmDeleteProps> = ({
         </GenericButton>
         <GenericButton
           type="button"
-          variant="error"
+          variant="primary"
           onClick={onConfirm}
           disabled={loading}
         >
