+++ Frontend/tkd_hub_web/src/app/components/users/EditUser.tsx
@@ -1,7 +1,5 @@
 import React, { useState, useEffect } from "react";
-import LabeledInput from "@/app/components/common/inputs/LabeledInput";
-import { GenericSelector } from "@/app/components/common/selectors/GenericSelector";
-import FormActionButtons from "@/app/components/common/actionButtons/FormActionButtons";
+import EditUserForm from "./EditUserForm";
 import { EditModal } from "@/app/components/common/modals/EditModal";
 
 type EditUserProps = {
@@ -9,28 +7,18 @@ type EditUserProps = {
   onClose: (refreshList?: boolean) => void;
 };
 
-type User = {
-  id?: number;
-  name: string;
-  email: string;
-  role: string;
-};
-
-const defaultUser: User = {
-  name: "",
-  email: "",
-  role: "",
-};
-
-const roles = [
-  { id: 1, value: "Admin", label: "Admin" },
-  { id: 2, value: "User", label: "User" },
-];
-
 export default function EditUser({ userId, onClose }: EditUserProps) {
-  const [user, setUser] = useState<User>(defaultUser);
   const [loading, setLoading] = useState(false);
-  const [saving, setSaving] = useState(false);
+  const [user, setUser] = useState<{
+    id?: number;
+    name: string;
+    email: string;
+    role: string;
+  }>({
+    name: "",
+    email: "",
+    role: "",
+  });
 
   useEffect(() => {
     if (userId) {
@@ -46,28 +34,9 @@ export default function EditUser({ userId, onClose }: EditUserProps) {
         setLoading(false);
       }, 500);
     } else {
-      setUser(defaultUser);
+      setUser({ name: "", email: "", role: "" });
     }
   }, [userId]);
-
-  function handleChange(e: React.ChangeEvent<HTMLInputElement>) {
-    setUser({ ...user, [e.target.name]: e.target.value });
-  }
-
-  function handleRoleChange(id: number | null) {
-    const selected = roles.find((r) => r.id === id);
-    setUser({ ...user, role: selected ? selected.value : "" });
-  }
-
-  function handleSubmit(e: React.FormEvent) {
-    e.preventDefault();
-    setSaving(true);
-    // Replace with real save logic
-    setTimeout(() => {
-      setSaving(false);
-      onClose(true);
-    }, 500);
-  }
 
   if (loading) {
     return <div className="text-center">Loading user...</div>;
@@ -77,44 +46,15 @@ export default function EditUser({ userId, onClose }: EditUserProps) {
     <EditModal
       open={true}
       title={userId ? "Edit User" : "Add User"}
-      saving={saving}
+      saving={false}
       onClose={onClose}
     >
-      <form onSubmit={handleSubmit} className="space-y-4">
-        <LabeledInput
-          label="Name"
-          name="name"
-          value={user.name}
-          onChange={handleChange}
-          required
-          disabled={saving}
-        />
-        <LabeledInput
-          label="Email"
-          name="email"
-          type="email"
-          value={user.email}
-          onChange={handleChange}
-          required
-          disabled={saving}
-        />
-        <GenericSelector
-          items={roles}
-          value={roles.find((r) => r.value === user.role)?.id ?? null}
-          onChange={handleRoleChange}
-          getLabel={(r) => r.label}
-          getId={(r) => r.id}
-          label="Role"
-          required
-          disabled={saving}
-          placeholder="Select role"
-        />
-        <FormActionButtons
-          onCancel={() => onClose()}
-          onSubmitLabel={userId ? "Save" : "Create"}
-          loading={saving}
-        />
-      </form>
+      <EditUserForm
+        user={user}
+        setUser={setUser}
+        userId={userId}
+        onClose={onClose}
+      />
     </EditModal>
   );
 }
